LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-nonascii-expected.txt

CONSOLE MESSAGE: The value for Content Security Policy directive 'script-src' contains an invalid character: '127.0.0.1/ßisnotSorB/'. Non-whitespace characters outside ASCII 0x21-0x7E must be percent-encoded, as described in RFC 3986, section 2.1: http://tools.ietf.org/html/rfc3986#section-2.1.
Sources containing non-ascii characters should be ignored, and should generate warnings.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
